+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `json_maker` function optimized
+- `dl_maker` function optimized
+- `tgf_maker` function optimized
+- `lp_maker` function optimized
 ## [0.5] - 2020-07-01
 ### Added
 - TSV format